refactor(runtime): extract prompt module from agent_loop
The pre-LLM-call prompt setup cluster moves into `agent_loop/prompt.rs`:

- `push_filtered_user_message` — PII-filter + sender-prefix wrap +
  push-to-session for the inbound user message. Handles both
  string-only and `Vec<ContentBlock>` (image-bearing) shapes, with
  the prefix landing on the first text block so privacy filtering
  runs over the raw input.
- `remember_interaction_best_effort` — async episodic-memory write.
  Tries embedding-aware persist first, falls back to plain remember
  on embedding failure, and just `warn!`s on persist errors —
  never blocks the agent loop.
- `proactive_item_to_fragment` — module-private; only used by
  `setup_recalled_memories`.
- Context / setup structs (all `pub(super)` with `pub(super)` fields
  so main loop and end_turn can construct them):
  `PromptExperimentSelection`, `RecallSetup`, `RecallSetupContext`,
  `PromptSetup`, `PromptSetupContext`, `PreparedMessages`.
- `reply_directives_from_parsed` — used by both end_turn.rs and the
  main-loop silent-result builder; `pub(super)`.
- `select_running_experiment` — A/B experiment resolution from
  manifest + per-session metadata.
- `setup_recalled_memories` — fan-in of `auto_retrieve` + proactive
  store + ACL-aware filter (RBAC M3, #3054).
- `build_prompt_setup` — builds the system prompt + memory-context
  message that gets prepended.
- `prepare_llm_messages` — runs session repair + safe trim before
  the LLM call, surfacing repair stats for the next helper.
- `log_repair_stats` — single-site logger for the repair stats.

end_turn.rs picks up its remaining cross-module deps from prompt
(`reply_directives_from_parsed`, `remember_interaction_best_effort`).
mod.rs imports the entire prompt API surface via
`use self::prompt::{...}`. mod tests in mod.rs already accesses
everything via `use super::*` which now resolves transitively
through the `pub(super)` items.

Behavior-preserving. mod.rs: 10,893 -> 10,284 LOC (-609).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
